Correct frandmod signedness contract to the retail-proven s32 form
The retail DOL converts frandmod results to float through the signed
xoris 0x8000 / 4330000080000000 sequence in CharModelLandDustCreate
(0x8005FAA8-0x8005FAC4), while charman.c's preprocessed TU contains no
frandmod declaration at all: the implicit C89 int contract, not the
shared header, produced those bytes. TUs that do include the header
(actman.c, colman.c) match retail only with unsigned conversions and
cmplwi zero compares at their frandmod use sites.

Flip include/game/frand.h and src/game/frand.c to s32 frandmod(s32
modulus), matching the dol-authority contract already recorded for
mdpartydll/stage.c in config/GP6E01/tu_declarations.json, and preserve
the proven unsigned sites in actman.c/colman.c with explicit (u32)
casts. frand.o is byte-identical either way because the body already
divides unsigned ((frand_seed & 0x7FFFFFFF) % (u32)modulus, divwu at
0x8003FD58).

Full container gate after the change: dtk shasum -c
config/GP6E01/build.sha1 reports 137 files OK, main.dol byte-identical
to retail (b897e6ade6b3a0cd2f9907689f38a3b19c327e70).

Evidence: docs/frandmod_signedness_evidence.md

+++ src/game/colman.c
@@ -2571,8 +2571,8 @@ static int _BodyColNarrow(void)
             }
             VECSubtract(&col2.point, &col1.point, &colDelta);
             if(VECSquareMag(&colDelta) < 0.0001f) {
-                colDelta.x = (frandmod(20)-10.0f)*0.01f;
-                colDelta.z = (frandmod(20)-10.0f)*0.01f;
+                colDelta.x = ((u32)frandmod(20)-10.0f)*0.01f;
+                colDelta.z = ((u32)frandmod(20)-10.0f)*0.01f;
                 colDelta.y = 0.01f;
             }
             VECNormalize(&colDelta, &colDeltaNorm);
@@ -2838,7 +2838,7 @@ static BOOL _BodyApplyColAttr(COLBODY *bodyP, COL_ATTRPARAM *attrParam, int code
         bodyP->moveDir.x = 0;
         bodyP->moveDir.y = 0;
         bodyP->moveDir.z = 0;
-        VECScale(&point->normal, &bounce, bodyP->param.bounce*(0.3f+(frandmod(70)/100.0f)));
+        VECScale(&point->normal, &bounce, bodyP->param.bounce*(0.3f+((u32)frandmod(70)/100.0f)));
         VECScale(&bounce, &bounce, bodyP->param.radius);
         VECAdd(&bodyP->pos[posIdx], &bounce, &bodyP->pos[posIdx]);
         bodyP->oldPos = bodyP->pos[posIdx];
